requirements.txt
python-dateutil
setuptools_rust
django==3.1.5
selenium==3.141.0
django_compressor==2.4
django-libsass==0.8
djangorestframework==3.12.2
django-registration==3.1.1
django-extensions==3.1.1
django-dbbackup==3.3.0
whitenoise==5.2.0
lxml==4.6.2
beautifulsoup4==4.9.3
requests
retrying
pysocks
markdown
daphne
psycopg2-binary
tqdm